use an F instead of static_cast

+++ include/RE/N/NiPoint3.cpp
@@ -46,7 +46,7 @@ namespace RE
 
 	NiPoint3 NiPoint3::operator/(float a_scalar) const
 	{
-		return operator*(static_cast<float>(1.0) / a_scalar);
+		return operator*(1.0F / a_scalar);
 	}
 
 	NiPoint3 NiPoint3::operator-() const
@@ -80,7 +80,7 @@ namespace RE
 
 	NiPoint3& NiPoint3::operator/=(float a_scalar)
 	{
-		return operator*=(static_cast<float>(1.0) / a_scalar);
+		return operator*=(1.0F / a_scalar);
 	}
 
 	NiPoint3 NiPoint3::Cross(const NiPoint3& a_pt) const
